// types/talep.ts
export interface Talep {
  id: string
  talepSahibi: string
  talepSahibiAciklamasi: string
  talepSahibiDigerAciklama?: string
  talepIlcesi: string
  bolge: string
  hatNo: string
  isletici: string
  talepOzeti: string
  talepIletimSekli: string
  evrakTarihi?: string
  evrakSayisi?: string
  yapılanIs: string
  talepDurumu: string
  guncellemeTarihi: string
}